Moved malaria api code to its own Django app
